clearing #1 atlas->array

+++ project/inc/components/parenthood.hpp
@@ -4,9 +4,12 @@
 namespace PARENTHOOD {
 
     struct Base {
-        u64 childrenCount = 0;
+        u16 childrenCount = 0;
         GameObjectID* children = nullptr;
     };
+
+    // ChildrenCollection = 2
+    // size, parent( size, child, child, child), parent(size, child, child)
 
     struct Parenthood {
         GameObjectID id = 0;
